+++ .github/workflows/run-unittest.yml
@@ -34,6 +34,9 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-    - name: Test with unittest
+    - name: Test with Unit Test
       run: |
-        python3 -m unittest -v
+        python3 -m unittest discover -s tests/unit -v
+    - name: Test with Integration Test
+      run: |
+        python3 -m unittest discover -s tests/integration -v
