+++ tests/unit/test_SA_r1_Rat24S.py
@@ -1,9 +1,9 @@
 import unittest
 import os
-from tests.helpers import write_to_file, get_result_from_parser
+from tests.unit.helpers import write_to_file, get_result_from_parser
 
 
-class ReturnTestCase(unittest.TestCase):
+class R1_Rat24STestCase(unittest.TestCase):
     SAMPLE_FILE_PATH = "tests/sample1.txt"
 
     def setUp(self):
@@ -14,19 +14,9 @@ class ReturnTestCase(unittest.TestCase):
         if os.path.exists(self.SAMPLE_FILE_PATH):
             os.remove(self.SAMPLE_FILE_PATH)
 
-    def test_if_no_else(self):
-        input_string = "$ $ $ if (a == b ) return c; endif $"
-        expected_output = True
-
-        # Act
-        write_to_file(self.SAMPLE_FILE_PATH, input_string)
-        actual_output = get_result_from_parser(self.SAMPLE_FILE_PATH)
-
-        # Assert
-        self.assertEqual(actual_output, expected_output)
-    
-    def test_if_else(self):
-        input_string = "$ $ $ if (a == b ) return c; else return 85; endif $"
+    def test_one_function_definition(self):
+        # Arrange
+        input_string = "$ function f1() { a = 1; } $ $ a = b; $"
         expected_output = True
 
         # Act
@@ -36,20 +26,10 @@ class ReturnTestCase(unittest.TestCase):
         # Assert
         self.assertEqual(actual_output, expected_output)
 
-    def test_if_no_if(self):
-        input_string = "$ $ $ (a == b ) return c; endif $"
-        expected_output = False
-
-        # Act
-        write_to_file(self.SAMPLE_FILE_PATH, input_string)
-        actual_output = get_result_from_parser(self.SAMPLE_FILE_PATH)
-
-        # Assert
-        self.assertEqual(actual_output, expected_output)
-    
-    def test_if_no_endif(self):
-        input_string = "$ $ $ if (a == b ) return c; $"
-        expected_output = False
+    def test_no_function_definition(self):
+        # Arrange
+        input_string = "$ $ $ a = b; $"
+        expected_output = True
 
         # Act
         write_to_file(self.SAMPLE_FILE_PATH, input_string)
@@ -58,8 +38,21 @@ class ReturnTestCase(unittest.TestCase):
         # Assert
         self.assertEqual(actual_output, expected_output)
 
-    def test_if_no_open_paren(self):
-        input_string = "$ $ $ if a == b ) return c; endif $"
+    def test_no_declarations_no_function_definitions(self):
+        # Arrange
+        input_string = "$ $ $ num = 1; $"
+        expected_output = True
+
+        # Act
+        write_to_file(self.SAMPLE_FILE_PATH, input_string)
+        actual_output = get_result_from_parser(self.SAMPLE_FILE_PATH)
+
+        # Assert
+        self.assertEqual(actual_output, expected_output)
+
+    def test_not_end_with_dollar_symbol(self):
+        # Arrange
+        input_string = "$ $ $ a = b; $xyz"
         expected_output = False
 
         # Act
@@ -69,8 +62,9 @@ class ReturnTestCase(unittest.TestCase):
         # Assert
         self.assertEqual(actual_output, expected_output)
 
-    def test_if_no_close_paren(self):
-        input_string = "$ $ $ if (a == b  return c; endif $"
+    def test_the_input_is_empty(self):
+        # Arrange
+        input_string = ""
         expected_output = False
 
         # Act
@@ -80,4 +74,6 @@ class ReturnTestCase(unittest.TestCase):
         # Assert
         self.assertEqual(actual_output, expected_output)
 
-    
+
+if __name__ == '__main__':
+    unittest.main()
