+++ tests/unit/test_SA_r21_scan.py
@@ -1,6 +1,6 @@
 import unittest
 import os
-from tests.helpers import write_to_file, get_result_from_parser
+from tests.unit.helpers import write_to_file, get_result_from_parser
 
 class PrintTestCase(unittest.TestCase):
     SAMPLE_FILE_PATH = "tests/sample1.txt"
